fix(compiler): eliminate 100% of compiler and linker warnings in Swift and C11
- Fix Mach-O common section over-alignment by initializing s_length_slot and s_offset_slot
- Fix Swift Sendable closure capture and Codable conformances in TTZipBench/TTZipCore
- Fix signed/unsigned comparisons in lzfse and fast-lzma2
- Silence unused-but-set variables in zopfli and snappy in Release mode

+++ Sources/CTTZipBridge/lzfse/lzfse_decode_base.c
@@ -237,7 +237,7 @@ static int lzfse_decode_lmd(lzfse_decoder_state *s) {
       if (D >= 8 || D >= M)
         copy(dst, dst - D, M);
       else
-        for (size_t i = 0; i < M; i++)
+        for (int32_t i = 0; i < M; i++)
           dst[i] = dst[i - D];
       dst += M;
     }
@@ -253,7 +253,7 @@ static int lzfse_decode_lmd(lzfse_decoder_state *s) {
       //  or there isn't; if there is, we copy the whole thing and
       //  update all the pointers and lengths to reflect the copy.
       if (L <= remaining_bytes) {
-        for (size_t i = 0; i < L; i++)
+        for (int32_t i = 0; i < L; i++)
           dst[i] = lit[i];
         dst += L;
         lit += L;
@@ -265,7 +265,7 @@ static int lzfse_decode_lmd(lzfse_decoder_state *s) {
       //  L, and report that the destination buffer is full. Note that
       //  we always write right up to the end of the destination buffer.
       else {
-        for (size_t i = 0; i < remaining_bytes; i++)
+        for (ptrdiff_t i = 0; i < remaining_bytes; i++)
           dst[i] = lit[i];
         dst += remaining_bytes;
         lit += remaining_bytes;
@@ -277,7 +277,7 @@ static int lzfse_decode_lmd(lzfse_decoder_state *s) {
       //  before finishing, we return to the caller indicating that
       //  the buffer is full.
       if (M <= remaining_bytes) {
-        for (size_t i = 0; i < M; i++)
+        for (int32_t i = 0; i < M; i++)
           dst[i] = dst[i - D];
         dst += M;
         remaining_bytes -= M;
@@ -291,7 +291,7 @@ static int lzfse_decode_lmd(lzfse_decoder_state *s) {
                  //
                  // But we still set M = 0, to maintain the post-condition.
       } else {
-        for (size_t i = 0; i < remaining_bytes; i++)
+        for (ptrdiff_t i = 0; i < remaining_bytes; i++)
           dst[i] = dst[i - D];
         dst += remaining_bytes;
         M -= remaining_bytes;
